[FIX] account, auth_signup, sale, website_sale: settings robustness
- Delete the tax corresponding to the default 'Sales Tax'
- Access the Accounting settings

Error since record could not be found.

In the case of an `ir.config_parameter` referring to a record ID
(i.e. Many2one), the deletion of this record is not propagated to the
associated `ir.config_parameter`. Therefore, the system tries to access
a deleted record without fallback.

It's not really possible to solve it globally in a stable version, but
we can manually handle these cases by providing a fallback (usually
return `False`).

opw-781864