[FIX] stock: Fix strange ORM bidouille on picking creation
Purpose
=======

This bidouille on the ORM command seems to be from another age where the onchange on o2m list were not working well.

The issue is that _convert_to_write could return commands like (6, 0, []).

In that case the last tuple element is not dictionary, but a list.

Specification
=============

Fix that brol.