[FIX] point_of_sale: some events changed to jquery
Before this commit, it was not possible to type on a physical keyboard
the amount of a payment in the relevant pos screen

This was due to the refactoring of BarcodeEvents (commit d7d2d17283ccfd345e260f027a74abcab7e30a0f)
which converted vanilla JS events to Jquery ones, which the pos was not ready for.

After this commit, it is possible to type the amount of a payment
in the payment screen, just as it worked in 10.0 so far.

OPW 781242

closes #21094